refactor: create handleRefineOptions
