+++ packages/core/src/definitions/helpers/index.ts
@@ -6,3 +6,4 @@ export { queryKeys } from "./queryKeys";
 export { hasPermission } from "./hasPermission";
 export { routeGenerator } from "./routeGenerator";
 export { createTreeView } from "./treeView/createTreeView";
+export { handleRefineOptions } from "./handleRefineOptions";
